% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altools_general_2020.R
\name{at_count_words}
\alias{at_count_words}
\title{Count the number of words in a string.}
\usage{
at_count_words(x)
}
\arguments{
\item{x}{string or vector of stings (in mutate case) to count words (or spaces in reality)}

\item{...}{Not used}
}
\description{
This actually counts the number of spaces so be sure to stringr::str_trim your whitespaces out from start and end. Check out
\href{https://stat.ethz.ch/R-manual/R-devel/library/base/html/regex.html}{details on s and other regular expressions}
}
\examples{
at_count_words("test this little thing")
at_count_words(" watch out for white spaces ! ")
}
