% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altools_time.R
\name{at_getSeason}
\alias{at_getSeason}
\title{Assign dates to a season}
\usage{
at_getSeason(dates, ...)
}
\arguments{
\item{dates}{Dates to find season for}

\item{...}{Not used}
}
\description{
Convert dates from any year to 2012 dates (leap year) and asssign a season. Stole it from
\href{https://stackoverflow.com/users/980833/josh-obrien}{Josh O'Brien} from this
\href{https://stackoverflow.com/questions/9500114/find-which-season-a-particular-date-belongs-to}{post}
}
\examples{
my.dates <- as.Date("2011-12-01", format = "\%Y-\%m-\%d") + 0:60
head(at_getSeason(my.dates), 24)
}
