% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altools_general_2020.R
\name{at_na_remove}
\alias{at_na_remove}
\title{remove NAs.  Works to remove a na from a list to if they are a list element or are within list elements}
\usage{
at_na_remove(x)
}
\arguments{
\item{x}{a list that needs NAs removed from}
}
\description{
remove NAs. Needed this when NA was one of the unique variables in
a column that got turned into a list
}
