% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altools_general_2020.R
\name{at_na_remove_column}
\alias{at_na_remove_column}
\title{remove columns that have only NAs from dataframe. Pipeable.}
\usage{
at_na_remove_column(df)
}
\arguments{
\item{df}{dataframe}

\item{...}{Not used. Can be used to pass parameters to other functions.}
}
\description{
Function came from
\href{https://stackoverflow.com/questions/15968494/how-to-delete-columns-that-contain-only-nas}{here}
}
