% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altools_general_2020.R
\name{at_remove_columns_named_this}
\alias{at_remove_columns_named_this}
\title{Remove columns with a specific string anywhere in the name.}
\usage{
at_remove_columns_named_this(df, string, ...)
}
\arguments{
\item{df}{dataframe}

\item{string}{Character string in quotations contained within columns to be removed from dataframe}

\item{...}{Not used. Can be used to pass parameters to other functions.}
}
\description{
Function came from
\href{https://stackoverflow.com/questions/41815039/remove-columns-that-contain-a-specific-word}{here}
}
