% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altools_time.R
\name{at_time_excel_number}
\alias{at_time_excel_number}
\title{Converts an excel numeric representation of a date to YY-mm-dd.}
\usage{
at_time_excel_number(date)
}
\arguments{
\item{date}{Date to convert from number to posix}

\item{...}{Not used}
}
\description{
Converts an excel numeric representation of a date to YY-mm-dd. Got it from this Breza
\href{https://stackoverflow.com/questions/43230470/how-to-convert-excel-date-format-to-proper-date-with-lubridate}{kid}
}
\examples{
as.Date(42705, origin = "1899-12-30")
}
