% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altools_spreadsheets.R
\name{at_trim_xlsheet}
\alias{at_trim_xlsheet}
\title{Clean up excel spreadsheet imports}
\usage{
at_trim_xlsheet(df, column_last = ncol(df), row_first = 1)
}
\arguments{
\item{df}{dataframe imported from excel that needs to be trimmed and cleaned}

\item{column_last}{numeric identifier for the last column to be imported}

\item{row_first}{numeric identifier for the first row to be imported}
}
\description{
Clean up excel spreadsheet imports by triming the number of columns, specifying which row should
be the top row
}
