% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altools_spreadsheets.R
\name{at_trim_xlsheet2}
\alias{at_trim_xlsheet2}
\title{Clean up excel spreadsheet imports}
\usage{
at_trim_xlsheet2(df, column_last = ncol(df))
}
\arguments{
\item{df}{dataframe imported from excel that needs to be trimmed and cleaned}

\item{column_last}{numeric identifier for the last column to be imported.  defaults to the last column present.}
}
\description{
Clean up excel spreadsheet imports by triming the number of columns, automatically chooses first complete row which row should
be the top row useing "complete.cases"
}
